pyproject.toml
[tool.poetry]
name = "exosky-nasa"
version = "0.1.0"
description = "This API provides endpoints to fetch and manage data on exoplanets and stars"
authors = ["Ihor Savenko <csigorek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
astroquery = "^0.4.7"
astropy = "^6.1.3"
numpy = "^2.1.1"
requests = "^2.32.3"
pandas = "^2.2.2"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
redis = "^5.0.8"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
pillow = "^10.4.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
